requirements.txt
Django==5.2.3
psycopg2_binary==2.9.10
python-dotenv==1.1.0
djangorestframework==3.16.0
django-filter==25.1
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.5.0
pytest-django==4.11.1